+ WIP: Cubemap's and skybox support, new section: drawSkybox in progress

+++ src/core/skinnedmesh.cpp
@@ -518,7 +518,7 @@ void SkinnedMesh::boneTransform(float timeInSeconds, std::vector<glm::mat4>& Tra
 
 	float TicksPerSecond = (float)(m_pScene->mAnimations[currentAnimation]->mTicksPerSecond != 0 ? m_pScene->mAnimations[currentAnimation]->mTicksPerSecond : 25.0f);
 	float TimeInTicks = timeInSeconds * TicksPerSecond;
-	float AnimationTime = fmod(TimeInTicks, animDuration);
+	float AnimationTime = (float)fmod(TimeInTicks, animDuration);
 
 	ReadNodeHeirarchy(AnimationTime, m_pScene->mRootNode, Identity);
